requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
discord==1.7.3
discord.py==1.7.3
et-xmlfile==1.1.0
google-api-core==2.0.1
google-api-python-client==2.20.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gspread==4.0.1
httplib2==0.19.1
idna==3.2
multidict==5.1.0
mustache==0.1.4
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.9
pandas==1.5.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==2.4.7
pystache==0.6.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==6.0.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.6
wget==3.2
yarl==1.6.3
